feat: extract shared API types to shared crate
