+++ server/Cargo.toml
@@ -13,6 +13,7 @@ dioxus-router = "0.7"
 dioxus-ssr = "0.7"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+omnibus-shared = { path = "../shared" }
 sqlx = { version = "0.8", features = ["runtime-tokio-rustls", "sqlite", "macros"] }
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
