+++ shared/src/lib.rs
@@ -3,3 +3,34 @@
 //! Keep this crate free of Dioxus and transport-layer dependencies so both
 //! `#[server]` functions (web) and `reqwest` calls (mobile) can depend on it
 //! without dragging in platform-specific trees.
+
+use serde::{Deserialize, Serialize};
+
+/// User-configurable paths for the ebook and audiobook libraries.
+#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
+pub struct Settings {
+    pub ebook_library_path: Option<String>,
+    pub audiobook_library_path: Option<String>,
+}
+
+/// Response payload for the counter endpoints (`GET /api/value`,
+/// `POST /api/value/increment`).
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct ValueResponse {
+    pub value: i64,
+}
+
+/// One half of the library listing (either ebooks or audiobooks).
+#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
+pub struct LibrarySection {
+    pub path: Option<String>,
+    pub files: Vec<String>,
+    pub error: Option<String>,
+}
+
+/// Response payload for `GET /api/library`.
+#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
+pub struct LibraryContents {
+    pub ebooks: LibrarySection,
+    pub audiobooks: LibrarySection,
+}
